Consolidate the example app code
